+++ packages/ramps-controller/src/index.ts
@@ -138,6 +138,7 @@ export { createRequestSelector } from './selectors';
 export type {
   TransakServiceActions,
   TransakServiceEvents,
+  TransakServiceOrderUpdateEvent,
   TransakServiceMessenger,
   TransakAccessToken,
   TransakUserDetails,
@@ -160,6 +161,9 @@ export type {
   TransakUserLimits,
   TransakIdProofStatus,
   PatchUserRequestBody as TransakPatchUserRequestBody,
+  PusherFactory,
+  PusherLike,
+  ChannelLike,
 } from './TransakService';
 export {
   TransakApiError,
@@ -178,4 +182,7 @@ export type {
   TransakServiceGetOrderAction,
   TransakServiceRequestOttAction,
   TransakServiceGeneratePaymentWidgetUrlAction,
+  TransakServiceSubscribeToOrderAction,
+  TransakServiceUnsubscribeFromOrderAction,
+  TransakServiceDisconnectWebSocketAction,
 } from './TransakService-method-action-types';